Solution for assignment 3 - a static web page with a drop down menu that only appears on a phone-size screen and using bootstrap for most of the styling.
There were some issues with this solution.  For a while the current
version of bootstrap (3.3.7) did not work for the collapsing menu.
Substituting in v3.3.6 worked and subsequently 3.3.7 worked (no HTML or
CSS changes were made).  Also, the bootstrap row element is not working
well for screen widths below 768px.  Below that the row has no margins.